+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/adapter/input/data/response/CrowdfundingListResponse.kt
@@ -1,13 +1,12 @@
 package com.project.indistraw.crowdfunding.adapter.input.data.response
 
 import com.project.indistraw.crowdfunding.domain.StatusType
-import java.math.BigDecimal
 
 data class CrowdfundingListResponse(
     val idx: Long,
     val title: String,
     val description: String,
-    val percentage: BigDecimal,
+    val percentage: Double,
     val thumbnailUrl: String,
     val status: StatusType
 )
