+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/adapter/input/data/response/RewardResponse.kt
@@ -1,12 +1,10 @@
 package com.project.indistraw.crowdfunding.adapter.input.data.response
 
-import java.math.BigDecimal
-
 data class RewardResponse(
     val idx: Long,
     val title: String,
     val description: String,
-    val price: BigDecimal,
+    val price: Long,
     val imageUrl: String,
     val totalCount: Long?
 )
