+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/domain/Amount.kt
@@ -1,10 +1,9 @@
 package com.project.indistraw.crowdfunding.domain
 
 import com.project.indistraw.crowdfunding.application.common.annotation.Aggregate
-import java.math.BigDecimal
 
 @Aggregate
 data class Amount(
-    val totalAmount: BigDecimal,
-    val targetAmount: BigDecimal
+    val totalAmount: Long,
+    val targetAmount: Long
 )
